+++ modules/DOMUtils.js
@@ -43,3 +43,12 @@ export const supportsPopStateOnHashChange = () =>
  */
 export const supportsGoWithoutReloadUsingHash = () =>
   window.navigator.userAgent.indexOf('Firefox') === -1
+
+/**
+ * Returns true if a given popstate event is an extraneous WebKit event.
+ * Accounts for the fact that Chrome on iOS fires real popstate events
+ * containing undefined state when pressing the back button.
+ */
+export const isExtraneousPopstateEvent = event =>
+  event.state === undefined &&
+  navigator.userAgent.indexOf('CriOS') === -1