+++ CHANGES.md
@@ -1,3 +1,7 @@
+## HEAD
+
+- Fix the back button on Chrome iOS
+
 ## [v4.4.0]
 > Nov 1, 2016